geopmdpy/requirements.txt
build>=0.9.0
cffi>=1.14.5
dasbus>=1.6
docstring_parser>=0.13
jsonschema>=3.2.0
psutil>=5.8.0
pyyaml>=6.0.0
setuptools>=59.6.0
setuptools-scm>=6.4.2
prometheus_client[stats]>=0.9.0
seaborn[stats]>=0.11.2
protobuf>=3.12.4
grpcio>=1.30.2
defusedxml>=0.7.1
sdnotify>=0.3.1
